+++ modules/core/petra/petra-url-pattern-mapper/src/main/java/com/liferay/petra/url/pattern/mapper/factory/URLPatternMapperFactory.java
@@ -12,7 +12,11 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.petra.url.pattern.mapper.factory;
+
+import com.liferay.petra.url.pattern.mapper.URLPatternMapper;
+import com.liferay.petra.url.pattern.mapper.trie.DynamicSizeTrieURLPatternMapper;
+import com.liferay.petra.url.pattern.mapper.trie.StaticSizeTrieURLPatternMapper;
 
 import java.util.Map;
