+++ modules/core/petra/petra-url-pattern-mapper/src/main/java/com/liferay/petra/url/pattern/mapper/simple/SimpleURLPatternMapper.java
@@ -12,7 +12,9 @@
  * details.
  */
 
-package com.liferay.portal.remote.cors.internal.url.pattern.mapper;
+package com.liferay.petra.url.pattern.mapper.simple;
+
+import com.liferay.petra.url.pattern.mapper.base.BaseURLPatternMapper;
 
 import java.util.HashMap;
 import java.util.Map;
